requirements.txt
streamlit
pandas
gspread
oauth2client
plotly
prophet
wordcloud
matplotlib
google-generativeai